Remove RTX specific types
Replace use of osRtxMutex_t with the generic
mbed_rtos_storage_mutex_t. The type osRtxMutex_t is specific to RTX
and will not work with other back ends of cmsis-os2. By replacing it
with a generic mbed storage type, the proper space for a mutex will be
reserved by uvisor, regardless of the OS back end.